Add support for multi-packet responses
The server splits messages with a payload of more than 4096 byte into multiple
packets. However, the packet format does not support splitting, so we have to
send a second request with a different id and read response packets until we
get an answer to the "end marker" packet. This works because the server
processes all incoming messages strictly in order. [1]

Authentication packets still use only one packet as their response payload is
always empty.

The packet ids are now sequentially (starting from 1) generated, positive i32s.
Additionally a new public method get_id is added to the Packet struct because
the id was not accessible before.

[1]: https://developer.valvesoftware.com/wiki/Source_RCON_Protocol#Multiple-packet_Responses